+++ test/e2e/edge-async-local-storage/index.test.ts
@@ -1,64 +1,22 @@
-import { createNext } from 'e2e-utils'
-import { NextInstance } from 'e2e-utils'
+import { nextTestSetup } from 'e2e-utils'
 import { fetchViaHTTP } from 'next-test-utils'
 
 describe('edge api can use async local storage', () => {
-  let next: NextInstance
+  const { next } = nextTestSetup({
+    files: __dirname,
+  })
 
   const cases = [
     {
       title: 'a single instance',
-      code: `
-        export const config = { runtime: 'edge' }
-        const storage = new AsyncLocalStorage()
-  
-        export default async function handler(request) {
-          const id = request.headers.get('req-id')
-          return storage.run({ id }, async () => {
-            await getSomeData()
-            return Response.json(storage.getStore())
-          })
-        }
-  
-        async function getSomeData() {
-          try {
-            const response = await fetch('https://example.vercel.sh')
-            await response.text()
-          } finally {
-            return true
-          }
-        }
-      `,
-      expectResponse: (response, id) =>
+      route: '/api/single',
+      expectResponse: (response: any, id: string) =>
         expect(response).toMatchObject({ status: 200, json: { id } }),
     },
     {
       title: 'multiple instances',
-      code: `
-        export const config = { runtime: 'edge' }
-        const topStorage = new AsyncLocalStorage()
-  
-        export default async function handler(request) {
-          const id = request.headers.get('req-id')
-          return topStorage.run({ id }, async () => {
-            const nested = await getSomeData(id)
-            return Response.json({ ...nested, ...topStorage.getStore() })
-          })
-        }
-  
-        async function getSomeData(id) {
-          const nestedStorage = new AsyncLocalStorage()
-          return nestedStorage.run('nested-' + id, async () => {
-            try {
-              const response = await fetch('https://example.vercel.sh')
-              await response.text()
-            } finally {
-              return { nestedId: nestedStorage.getStore() }
-            }
-          })
-        }
-      `,
-      expectResponse: (response, id) =>
+      route: '/api/multiple',
+      expectResponse: (response: any, id: string) =>
         expect(response).toMatchObject({
           status: 200,
           json: { id: id, nestedId: `nested-${id}` },
@@ -66,40 +24,28 @@ describe('edge api can use async local storage', () => {
     },
   ]
 
-  afterEach(() => next.destroy())
-
   it.each(cases)(
-    'cans use $title per request',
-    async ({ code, expectResponse }) => {
-      next = await createNext({
-        files: {
-          'pages/index.js': `
-            export default function () { return <div>Hello, world!</div> }
-          `,
-          'pages/api/async.js': code,
-        },
-      })
+    'can use $title per request',
+    async ({ route, expectResponse }) => {
       const ids = Array.from({ length: 100 }, (_, i) => `req-${i}`)
 
       const responses = await Promise.all(
         ids.map((id) =>
-          fetchViaHTTP(
-            next.url,
-            '/api/async',
-            {},
-            { headers: { 'req-id': id } }
-          ).then((response) =>
-            response.headers.get('content-type')?.startsWith('application/json')
-              ? response.json().then((json) => ({
-                  status: response.status,
-                  json,
-                  text: null,
-                }))
-              : response.text().then((text) => ({
-                  status: response.status,
-                  json: null,
-                  text,
-                }))
+          fetchViaHTTP(next.url, route, {}, { headers: { 'req-id': id } }).then(
+            (response) =>
+              response.headers
+                .get('content-type')
+                ?.startsWith('application/json')
+                ? response.json().then((json) => ({
+                    status: response.status,
+                    json,
+                    text: null,
+                  }))
+                : response.text().then((text) => ({
+                    status: response.status,
+                    json: null,
+                    text,
+                  }))
           )
         )
       )
